+++ TDDToySample/app/src/main/java/com/duzi/tddtoysample/ui/MainActivity.kt
@@ -1,13 +1,23 @@
 package com.duzi.tddtoysample.ui
 
 import android.os.Bundle
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
 import com.duzi.tddtoysample.R
+import com.duzi.tddtoysample.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
 
+    private lateinit var binding: ActivityMainBinding
+
+     private val mainViewModel by viewModels<MainViewModel>()
+
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_main)
+        binding.lifecycleOwner = this@MainActivity
+        binding.viewModel = mainViewModel
     }
 }
